+++ China/Base.lproj/Main.storyboard
@@ -307,7 +307,21 @@
                     <view key="view" contentMode="scaleToFill" id="wzo-i8-BMj">
                         <rect key="frame" x="0.0" y="0.0" width="320" height="568"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <subviews>
+                            <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="nFE-X4-kUD">
+                                <rect key="frame" x="113" y="268" width="94" height="32"/>
+                                <fontDescription key="fontDescription" type="system" pointSize="16"/>
+                                <state key="normal" title="systemShare"/>
+                                <connections>
+                                    <action selector="systemShare:" destination="KaI-44-6m0" eventType="touchUpInside" id="ZJq-ue-HVC"/>
+                                </connections>
+                            </button>
+                        </subviews>
                         <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
+                        <constraints>
+                            <constraint firstItem="nFE-X4-kUD" firstAttribute="centerY" secondItem="wzo-i8-BMj" secondAttribute="centerY" id="HZf-v0-mun"/>
+                            <constraint firstItem="nFE-X4-kUD" firstAttribute="centerX" secondItem="wzo-i8-BMj" secondAttribute="centerX" id="XrQ-Fd-Gpq"/>
+                        </constraints>
                     </view>
                     <navigationItem key="navigationItem" title="System" id="PYc-XX-3SJ"/>
                 </viewController>
